requirements.txt
# Django
Django==5.2

# Django REST Framework
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-cors-headers==4.4.0

# API Documentation
drf-spectacular==0.27.2

# Database
mysqlclient==2.2.4

# Redis & Cache
django-redis==5.4.0
redis==5.0.8

# Utilities
python-decouple==3.8
Pillow==10.4.0
openpyxl==3.1.2

# Development
ipython==8.26.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
faker==33.1.0
